feat : implement LogOut function
